+++ packages/plugin-hardhat/src/index.ts
@@ -118,8 +118,8 @@ extendConfig((config: HardhatConfig) => {
   }
 });
 
-if (tryRequire('@nomiclabs/hardhat-etherscan')) {
-  subtask('verify:verify').setAction(async (args, hre, runSuper) => {
+if (tryRequire('@nomicfoundation/hardhat-verify')) {
+  subtask('verify:etherscan').setAction(async (args, hre, runSuper) => {
     const { verify } = await import('./verify-proxy');
     return await verify(args, hre, runSuper);
   });