+++ packages/plugin-hardhat/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+
+- Use `hardhat-verify` for proxy verification. ([#829](https://github.com/OpenZeppelin/openzeppelin-upgrades/pull/829))
+
 ## 2.0.0-alpha.0 (2023-06-20)
 
 - Use ethers v6 and hardhat-ethers v3. ([#817](https://github.com/OpenZeppelin/openzeppelin-upgrades/pull/817))